+++ src/common/web/integration/authorization/strategies/AuthorizationStrategies.ts
@@ -4,6 +4,7 @@ import { Service } from "../../../services/Service";
 import { AuthorizationStrategiesCatalog } from "./AuthorizationStrategiesCatalog";
 import { AuthorizationStrategy } from "./AuthorizationStrategy";
 import { BasicStrategy, createBasicStrategy } from "./basic/BasicStrategy";
+import { BearerStrategy, createBearerStrategy } from "./bearer/BearerStrategy";
 import { createOAuth2Strategy, OAuth2Strategy } from "./oauth2/OAuth2Strategy";
 
 /**
@@ -15,6 +16,7 @@ export function registerAuthorizationStrategies(): void {
     // New strategies go here
     AuthorizationStrategiesCatalog.registerItem(OAuth2Strategy.Strategy, createOAuth2Strategy);
     AuthorizationStrategiesCatalog.registerItem(BasicStrategy.Strategy, createBasicStrategy);
+    AuthorizationStrategiesCatalog.registerItem(BearerStrategy.Strategy, createBearerStrategy);
 
     // Print all available strategies for debugging purposes
     const names = Object.keys(AuthorizationStrategiesCatalog.items);