+++ src/common/web/ui/dialogs/authorization/BasicCredentialsDialog.vue
@@ -51,7 +51,7 @@ function getValidatorShape(): any {
         </div>
 
         <Fieldset legend="Credentials" class="r-form-fieldset">
-            <span v-if="!!dialogData.userData.config.user_id_optional" class="r-form-field">
+            <span v-if="!dialogData.userData.config.user_id_optional" class="r-form-field">
                 <IftaLabel>
                     <InputText name="name" v-model.trim="dialogData.userData.userName" fluid autofocus />
                     <label>{{ dialogData.userData.config.user_id_label }} <MandatoryMark /></label>
@@ -65,7 +65,7 @@ function getValidatorShape(): any {
                 :class="{ 'mt-5': !dialogData.userData.config.user_id_optional }"
             >
                 <IftaLabel>
-                    <Password name="password" v-model="dialogData.userData.userPassword" :feedback="false" fluid autofocus />
+                    <Password name="password" v-model.trim="dialogData.userData.userPassword" :feedback="false" toggle-mask fluid autofocus />
                     <label>{{ dialogData.userData.config.user_password_label }} <MandatoryMark /></label>
                 </IftaLabel>
                 <small>The {{ dialogData.userData.config.user_password_label.toLowerCase() }} for the external service.</small>