+++ src/connectors/base/integration/authorization/strategies/bearer/__init__.py
@@ -1,0 +1,1 @@
+from .bearer_configuration import get_bearer_strategy_configuration