+++ src/connectors/base/settings/__init__.py
@@ -1,7 +1,8 @@
 from .authorization_setting_ids import (
     AuthorizationSettingIDs,
-    OAuth2AuthorizationSettingIDs,
     BasicAuthorizationSettingIDs,
+    BearerAuthorizationSettingIDs,
+    OAuth2AuthorizationSettingIDs,
 )
 from .connector_setting_ids import ConnectorSettingIDs
 from .connector_settings import get_connector_settings