+++ src/connectors/inveniordm/metadata/__init__.py
@@ -1,8 +1,1 @@
-from .utils import (
-    parse_contributors,
-    parse_creators,
-    parse_dates,
-    parse_grants,
-    parse_subjects,
-)
 from .inveniordm_metadata import InvenioRDMMetadata, InvenioRDMMetadataCreator