+++ src/connectors/osf/execution/osf_job_executor.py
@@ -132,9 +132,13 @@ class OSFJobExecutor(ConnectorJobExecutor):
         exc: Exception,
         state_callbacks: ProjectExternalStateCallbacks,
     ) -> None:
-        if isinstance(exc, requests.exceptions.RequestException) and (
-            exc.response.status_code == HTTPStatus.NOT_FOUND
-            or exc.response.status_code == HTTPStatus.GONE
+        if (
+            isinstance(exc, requests.exceptions.RequestException)
+            and exc.response is not None
+            and (
+                exc.response.status_code == HTTPStatus.NOT_FOUND
+                or exc.response.status_code == HTTPStatus.GONE
+            )
         ):
             self.report_message(
                 "The previous project no longer exists, a new one will be created"