+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## v1.3.0 - xxxx-xx-xx
 
+### New
+
+- **datasafe** connector
+- Authorization strategy `bearer`
+
 ### Improvements
 
 - Made connection and transfer attempts use exponential delays to make them more error tolerant